Optimize combine_glyphs, add err checks
* combine_glyphs should not double-check hashset for values
* combine_glyphs doesn't need to have two extra iterations to compute min/max
* combine_glyphs should return None per docs on empty
* Make path params more generic
* Remove `freetype-rs` dependency - use (unpublished) one from sdf_glyph_renderer
* get rid of all panics (unwraps)
* add proper thiserror-based errors
* cleanup exports to be on the same top level
